made it use double as it make no difference, but is easier to use

+++ src/main/java/teamport/aether/helper/AetherMathHelper.java
@@ -24,7 +24,7 @@ public class AetherMathHelper {
         return -Math.log(1 - random.nextDouble());
     }
 
-    public static double distanceToSqr(int x, int y, int z, int x1, int y1, int z1) {
+    public static double distanceToSqr(double x, double y, double z, double x1, double y1, double z1) {
         double d3 = x - x1;
         double d4 = y - y1;
         double d5 = z - z1;
